+++ twill/extensions/csv_iterate.py
@@ -13,7 +13,7 @@ from twill import execute_file, log, namespaces
 __all__ = ['csv_iterate']
 
 
-def csv_iterate(filename, scriptname):
+def csv_iterate(file_name, script_name):
     """>> csv_iterate <csv_file> <script>
 
     For each line in <csv_file>, read in a list of comma-separated values,
@@ -21,10 +21,10 @@ def csv_iterate(filename, scriptname):
     """
     global_dict, local_dict = namespaces.get_twill_glocals()
 
-    reader = csv.reader(open(filename, "rb"))
+    reader = csv.reader(open(file_name))
     for i, row in enumerate(reader, 1):
-        log.debug('csv_iterate: on row %d of %s', i, filename)
+        log.debug('csv_iterate: on row %d of %s', i, file_name)
         for j, col in enumerate(row, 1):
-            global_dict["col%d" % (j,)] = col
+            global_dict[f"col{j}"] = col
 
-        execute_file(scriptname, no_reset=True)
+        execute_file(script_name, no_reset=True)